requirements.txt
python-dateutil~=2.8.2
pytz~=2023.3
PyYAML~=6.0
PyQt5~=5.15.9
h5py~=3.9.0
numpy~=1.25.0
scipy~=1.11.0
pyqtgraph~=0.13.3
lmfit~=1.2.1